requirements-examples.txt
scikit-image>=0.17.2
pydicom>=2.3.0
PyPDF2>=2.11.1
types-pytz>=2024.1.0.20240203
highdicom>=0.18.2
SimpleITK>=2.0.0
Pillow>=8.4.0
nibabel>=3.2.1
numpy-stl>=2.12.0
trimesh>=3.8.11
torch>=2.6.0
monai>=1.3.0